Import of KARASIK/Prima-1.29_01 from CPAN.
gitpan-cpan-distribution: Prima
gitpan-cpan-version:      1.29_01
gitpan-cpan-path:         KARASIK/Prima-1.29_01.tar.gz
gitpan-cpan-author:       KARASIK
gitpan-cpan-maturity:     developer

+++ AbstractMenu.c
@@ -23,7 +23,7 @@
  * OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF
  * SUCH DAMAGE.
  *
- * $Id: AbstractMenu.c,v 1.60 2007/11/11 12:48:07 dk Exp $
+ * $Id: AbstractMenu.c,v 1.61 2011/03/26 20:59:19 dk Exp $
  */
 
 #include "apricot.h"
@@ -448,7 +448,7 @@ new_av(  PMenuItemReg m, int level)
 	    SV * sv;
             int shift = ( m-> flags. checked ? 1 : 0) + ( m-> flags. disabled ? 1 : 0);
 	    if ( shift > 0) { /* has flags */
-	       int len = strlen( m-> variable);
+	       int len = (int) strlen( m-> variable);
                char * name = allocs( len + shift);
 	       if ( name) {
 		  int slen = len + shift;
